+++ include/delfem2/opengl/old/mshuni.cpp
@@ -843,9 +843,9 @@ DFM2_INLINE void delfem2::opengl::DrawMeshTri2D_Face(
 
 DFM2_INLINE void delfem2::opengl::DrawMeshTri2D_FaceDisp2D(
     const double *aXY,
-    unsigned int nXY,
+    size_t nXY,
     const unsigned int *aTri,
-    unsigned int nTri,
+    size_t nTri,
     const double *aDisp,
     int nstride_disp) {
   namespace lcl = ::delfem2::opengl::old::mshuni;
@@ -912,9 +912,9 @@ DFM2_INLINE void delfem2::opengl::DrawMeshTri2D_Edge(
 
 DFM2_INLINE void delfem2::opengl::DrawMeshTri2D_FaceColor(
     const unsigned int *aTri,
-    unsigned int nTri,
+    size_t nTri,
     const double *aXY,
-    unsigned int nXY,
+    size_t nXY,
     const double *aColor) {
   ::glBegin(GL_TRIANGLES);
   for (unsigned int itri = 0; itri < nTri; itri++) {
